% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb.R
\name{nb.entropy}
\alias{nb.entropy}
\title{Estimate Newcomb-Benford Entropy based on Shannon's entropy}
\usage{
nb.entropy(x, digits = 1)
}
\arguments{
\item{x}{a vector, ignoring the sign}

\item{digits}{first or second digit; defaults to 1}
}
\value{
a single value showing the entropy as a base-2 logarithm
}
\description{
Estimate Newcomb-Benford Entropy based on Shannon's entropy
}
\examples{
nb.entropy(runif(100))
}
