% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sp500}
\alias{sp500}
\title{Dataset from the Standard's and Poor 500 equity index between 1928 and 2023}
\format{
A dataframe with 23931 days and 7 variables
\describe{
\item{Date}{Observation date (working days)}
\item{Open}{Openning value of the stock}
\item{High}{Highest value of the stock during the day}
\item{Low}{Lowest value of the stock during the day}
\item{Close}{Closing value of the stock}
\item{Adj.Close}{Adjusted closing value}
\item{Volume}{Number of stocks traded}
}
}
\source{
{yahoo finance}
}
\usage{
sp500
}
\description{
Contains seven columns
}
\examples{
# example code
data(sp500) # lazy loading
}
\keyword{datasets}
